Update postalignqc handling of samtools output
